src/requirements.txt
imageio
imageio-freeimage
jsonschema
mcap
mcap-ros2-support
numpy